Restructure files, add has-config util
