+++ api/src/redis/utils/redis-config-available.ts
@@ -1,0 +1,10 @@
+import { useEnv } from '../../env.js';
+
+/**
+ * Check if Redis configuration exists in the current project's environment configuration
+ */
+export const redisConfigAvailable = () => {
+	const env = useEnv();
+
+	return 'REDIS' in env || Object.keys(env).some((key) => key.startsWith('REDIS_'));
+};
